+++ src/main/java/sprt/ReduceData.java
@@ -3,61 +3,61 @@ package sprt;
 import java.io.Serializable;
 
 public class ReduceData implements Serializable {
-	private double[] variance;
-	private double[] cBeta;
-	private double[] ZScore;
-	private double[] theta1;
-	private double[] SPRT;
+	private float[] variance;
+	private float[] cBeta;
+	private float[] ZScore;
+	private float[] theta1;
+	private float[] SPRT;
 	private int[] SPRTActivationStatus;
 	private int[][] forecastedActivationStatus;
 
 	public ReduceData(Config config) {
-		this.variance = new double[config.contrasts.numContrasts];
-		this.cBeta = new double[config.contrasts.numContrasts];
-		this.ZScore = new double[config.contrasts.numContrasts];
-		this.theta1 = new double[config.contrasts.numContrasts];
-		this.SPRT = new double[config.contrasts.numContrasts];
+		this.variance = new float[config.contrasts.numContrasts];
+		this.cBeta = new float[config.contrasts.numContrasts];
+		this.ZScore = new float[config.contrasts.numContrasts];
+		this.theta1 = new float[config.contrasts.numContrasts];
+		this.SPRT = new float[config.contrasts.numContrasts];
 		this.SPRTActivationStatus = new int[config.contrasts.numContrasts];
 		// int[config.ROW][config.contrasts.numContrasts];
 	}
 
-	public double getVariance(int index) {
+	public float getVariance(int index) {
 		return this.variance[index];
 	}
 
-	public void setVariance(int index, double value) {
+	public void setVariance(int index, float value) {
 		this.variance[index] = value;
 	}
 
-	public double getCBeta(int index) {
+	public float getCBeta(int index) {
 		return this.cBeta[index];
 	}
 
-	public void setCBeta(int index, double value) {
+	public void setCBeta(int index, float value) {
 		this.cBeta[index] = value;
 	}
 
-	public double getZScore(int index) {
+	public float getZScore(int index) {
 		return this.ZScore[index];
 	}
 
-	public void setZScore(int index, double value) {
+	public void setZScore(int index, float value) {
 		this.ZScore[index] = value;
 	}
 
-	public double getTheta1(int index) {
+	public float getTheta1(int index) {
 		return this.theta1[index];
 	}
 
-	public void setTheta1(int index, double value) {
+	public void setTheta1(int index, float value) {
 		this.theta1[index] = value;
 	}
 
-	public double getSPRT(int index) {
+	public float getSPRT(int index) {
 		return this.SPRT[index];
 	}
 
-	public void setSPRT(int index, double value) {
+	public void setSPRT(int index, float value) {
 		this.SPRT[index] = value;
 	}
